Map 284 - esri elevation api (#102)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Changed the API call for elevation from Google Maps Elevation API to
Esri Elevation API within the Map Proxy Service. Added unit test.
Finalized offset Z calculation.
<!--- Describe your changes in detail -->

## Related GitHub Issue
N/A
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
MAP-284
<!-- e.g. CAR-123 -->

## Motivation and Context
Previously generated child maps contain elevations calculated by the
Bing Maps Elevation API which uses the ellipsoid earth model. This API
was replaced with the Google Maps Elevation API, which seemingly uses a
different earth model (possibly the geoid model). This caused some
discrepancies in the elevations when creating a computed lane in an old
child map. The original lane would have an elevation of 39 and the new
computed lane would have an elevation of 72, which, ultimately, resulted
in errors when calculating the offsetZ
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Unit Tests and Local Deployment Tests
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ x ] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ x ] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ x ] I have added tests to cover my changes.
- [ x ] All new and existing tests passed.