Map159 movements container time restrictions (#141)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Added time restrictions to the approach pop up as a table of rows
containing three columns
- first column - radio button for MAP approach
- second column- approach type
- third column - time restrictions
If a map file is loaded, there will be one row with current approach and
time restrictions is set to none
If a row is selected for an approach, that is the one that will be used
for the MAP message encoding. Otherwise, if RGA is enabled, the entire
table will be encoded.
If a row is selected for an approach, the delete button is disabled. 
Time restrictions are disabled if rga is disabled. 
<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
MAP-159
<!-- e.g. CAR-123 -->

## Motivation and Context
Implement time restrictions for approaches: Currently, the ISD tool does
not support this feature. The time restrictions for the direction of
travel can be implemented within the
“fedgov-cv-ISDcreator-webapp/src/main/webapp/index.html” in the
“Approach Configuration” window’s “Approach Info” with “RGA only tag”.
An option to select days of the week and an option to select start/end
times need to be added so that the users can apply the time restrictions
for the direction of travel.
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Local deployment testing
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.