Warnings and Filtering for Speed Limits (#155)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Warnings and Filtering for Incompatible Speed Limits have been added to
the encoder.
<!--- Describe your changes in detail -->

## Related GitHub Issue
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
MAP-305
<!-- e.g. CAR-123 -->

## Motivation and Context
If a MAP Speed Limit Type cannot be converted to an RGA Speed Limit,
show the user a warning letting them know that it is not supported in
RGA yet and won’t be encoded. The warning should be shown for the
following speed limit types:
- Unknown   
- Vehicles With Trailers Min Speed   
- Vehicles With Trailers Max Speed   
- Vehicles With Trailers Night Max Speed  

If a RGA Speed Limit Type is not convertible to MAP, show the user a
warning letting them know that it is not supported in MAP and will not
be encoded. That warning should be shown for the following speed limit
types:
- Passenger Vehicles Max Speed   
- Passenger Vehicles Min Speed   

Additionally, these items should be handled:
- MAP should not show advisory speed limit choice.   
- Speed Limit Types Unknown, Vehicles with Trailers Min Speed, Vehicles
with Trailers Max Speed, Vehicles with Trailers Night Max Speed should
not show up in RGA.
- Speed Limit Types Passenger Vehicles Max Speed and Min Speed should
not show up in MAP
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Local Docker Deployment Testing
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.